Ensure Class element without Context does not crash compiler
